feat: Implement JavaScriptLoader for dynamic script execution and caching
- Added JavaScriptLoader class to load, compile, and execute JavaScript scripts using ClearScript.
- Implemented caching mechanism for compiled scripts to improve performance.
- Introduced methods for registering classes, global variables, functions, and events in the JavaScript environment.
- Added error handling and script evaluation capabilities.

feat: Create console command framework with command attributes

- Introduced CommandAttribute and BranchCommandAttribute for defining console commands and their metadata.
- Implemented ICommandHandler interface for command execution logic.
- Added ConsoleCommandProcessor to manage command registration and execution, supporting subcommands and branch commands.
- Developed HelpHandler to display available commands and their descriptions.

refactor: Enhance command processing and error handling

- Improved command parsing logic to support multiple subcommands and branch commands.
- Added detailed error messages for command execution failures.
- Implemented command information retrieval for help display.

chore: Add SubCommandAttribute for hierarchical command structure

- Introduced SubCommandAttribute to define subcommands with levels for better organization.

+++ Library/Applications/Servers/KaxServer/Models/StoreItem.cs
@@ -29,7 +29,7 @@ public class StoreItemPrice : IDataTable    // 商品的价格
     public int ParentId { get; set; }
     public string TableName => "StoreItemPrice";
     public string Title { get; set; } // 价格标题
-    public decimal Price { get; set; }  // 价格，这里是原价
+    public float Price { get; set; }  // 价格，这里是原价
     public float Rebate { get; set; } // 折扣率，0-1之间
     public float Duration { get; set; } // 时长
     public StoreItemDurationUnit DurationUnit { get; set; } // 时长单位
